+++ src/index.test.ts
@@ -1,5 +1,0 @@
-import sourcemap_support from 'source-map-support';
-
-sourcemap_support.install({
-	handleUncaughtExceptions: false,
-});