+++ src/paths.ts
@@ -35,6 +35,8 @@ export const LIB_DIR = `${LIB_DIRNAME}/`;
 export const CONFIG_SOURCE_PATH = 'gro.config.ts';
 export const CONFIG_BUILD_PATH = 'gro.config.js';
 
+export const MAIN_TEST_PATH = `${LIB_DIRNAME}/main.test.ts`;
+
 export const EXTERNALS_BUILD_DIRNAME = 'externals'; // TODO breaks the above trailing slash convention - revisit with trailing-slash branch
 export const EXTERNALS_BUILD_DIR_ROOT_PREFIX = `/${EXTERNALS_BUILD_DIRNAME}/`;